Refactor date format in working_time_test.exs and workingtime.ex
Refactor the date format in the working_time_test.exs and workingtime.ex files to improve consistency and readability.

Refactor database hostname configuration

Refactor the database hostname configuration to enhance flexibility and maintainability.

Add error handling in user routes

Add error handling in the user routes to improve the robustness and reliability of the application.

Add volume for persistent PostgreSQL data in docker-compose.yml

Add a volume configuration in the docker-compose.yml file to ensure persistent storage for PostgreSQL data.