refactor: Update Spline Viewer URL and add Tristant mode functionality
This commit refactors the code in App.vue to update the Spline Viewer URL based on the value of the isTristantMode variable. It also adds functionality to toggle Tristant mode by pressing Ctrl+T and clears the cache when the mode is toggled. Additionally, it adds the "tristant-mode" class to the body element when Tristant mode is enabled. The commit also includes a new image file "tristan.png" in the assets directory.